+++ src/components/obeliskForm/components/moving/moving.component.js
@@ -1,7 +1,8 @@
-import module from '../../../app.module';
-import { INTERVAL_DELAY } from '../../../config';
+import appModule from '../../../../app.module';
 
-import template from './moving.template.html';
+import { INTERVAL_DELAY } from '../../../../config';
+
+import template from './moving.component.html';
 
 class MovingController {
   constructor($interval, Obelisk) {
@@ -78,7 +79,7 @@ class MovingController {
   }
 }
 
-module.component('cgMoving', {
+appModule.component('cgMoving', {
   template: template,
   controller: MovingController
 });
