+++ src/components/obeliskForm/components/randomGeneration/randomGeneration.component.html
