+++ src/components/obeliskForm/components/randomLocation/randomLocation.component.html
