+++ src/components/obeliskForm/index.js
@@ -1,0 +1,10 @@
+import './obeliskForm.component';
+
+import './obelisk.service';
+
+import './components/randomGeneration/randomGeneration.component';
+import './components/moving/moving.component';
+
+import './components/nested/nested.component';
+import './components/randomLocation/randomLocation.component';
+import './components/elipticalMovement/elipticalMovement.component';
