+++ src/components/obeliskForm/obeliskForm.component.html
@@ -46,7 +46,6 @@
     </div>
   </div>
 
-
   <cg-random-generation></cg-random-generation>
   <cg-moving></cg-moving>
